move custom aplication download location to $custom_application_name
Prior to this commit the custom application git repo was cloned into the $custom_application_source_dir,
now it is cloned into $application_name.
This means that $custom_application_source_dir can now be used to specifiy a directory within the repo to point to the source.

Example: cb20test (github.com/ntb-ch/cb20)

The program cb20test is in the cb20 repo, but in a subdirectory: Software/cb20test.
Before this commit setting $custom_application_source_dir to "cb20test/Software/cb20test"
resulted in the repo being cloned into "cb20test/Software/cb20test", so the source was actually at
"cb20test/Software/cb20test/Software/cb20test". So cmake was still unable to find it, since it
looked only in "cb20test/Software/cb20test" (the root of the repo)

Now the repo is cloned into "cb20test" so "cb20test/Software/cb20test" is the source directory.
Setting $custom_application_source_dir to that will now result in a successful build, as cmake
is pointed to the right place.