Enable CMake to find EEROS when using SDK
When using cross compilation environment, CMake find_package()
function did not find the EEROS library. This happened since
the SDK script set the sysroot to the SDK folder. So CMake
did search at the wrong place. This commit ensures that the
libraries are installed in the SDK sysroot. CMake and other
applications will find the libraries in the default paths.